requirements.txt
cairo-lang
cairo-nile
openzeppelin-cairo-contracts
marshmallow-dataclass==8.5.3
black
pylint